move text generator into diary entry
