+++ spec/factories/diary_entries.rb
@@ -1,10 +1,5 @@
 FactoryGirl.define do
   factory :diary_entry do
-    heading "MyString"
-    introduction "MyString"
-    main_part "MyString"
-    closing "MyString"
-
     association :report
   end
 end
